requirements.txt
django
django-allauth[socialaccount]
google-api-python-client
google-auth-oauthlib
